Move PluginRepository to the common package
